+++ src/lib.rs
@@ -16,9 +16,9 @@
 //!
 //! # Overview
 //!
-//! `microbench` uses linear regression to estimate the execution time of code segments. For
-//! example, the following table might represent data collected by `microbench` about a code
-//! segment.
+//! `microbench` uses linear regression to estimate the execution time of code
+//! segments. For example, the following table might represent data collected by
+//! `microbench` about a code segment.
 //!
 //! | Iterations | Time (ns) |
 //! |------------|-----------|
@@ -28,10 +28,11 @@
 //! | 4          | 47        |
 //! | 5          | 56        |
 //!
-//! `microbench` of course takes many more than 5 samples and the number of iterations grows
-//! geometrically rather than linearly, but the idea remains the same. After collecting data like
-//! this, `microbench` uses ordinary least squares (OLS) linear regression to estimate the actual
-//! execution time of the code segment. Using OLS with the above data would yield an estimated
+//! `microbench` of course takes many more than 5 samples and the number of
+//! iterations grows geometrically rather than linearly, but the idea remains
+//! the same. After collecting data like this, `microbench` uses ordinary least
+//! squares (OLS) linear regression to estimate the actual execution time of the
+//! code segment. Using OLS with the above data would yield an estimated
 //! execution time of `9.6` nanoseconds with a goodness of fit (R²) of `0.992`.
 //!
 //! # Example
@@ -69,9 +70,6 @@
 
 #![warn(missing_copy_implementations, missing_debug_implementations, missing_docs)]
 
-#[cfg(feature="nightly")]
-extern crate test;
-
 mod utility;
 pub mod statistics;
 pub mod time;
@@ -80,92 +78,35 @@ use std::cmp;
 use std::mem;
 use std::time::{Duration};
 
+use crate::statistics::{Model};
 use crate::time::{Nanoseconds, Stopwatch};
+use crate::utility::{GeometricSequence, black_box, format_number};
 
-//================================================
-// Structs
-//================================================
-
-// Analysis ______________________________________
-
-/// A statistical analysis of a set of timing data.
-#[derive(Copy, Clone, Debug)]
-pub struct Analysis {
-    /// The y-intercept of the linear regression estimator.
-    pub alpha: Nanoseconds<f64>,
-    /// The slope of the linear regression estimator.
-    pub beta: Nanoseconds<f64>,
-    /// The goodness of fit of the linear regression estimator.
-    pub r2: f64,
-}
-
-// Bytes _________________________________________
+/// The maximum number of benchmark sample iterations.
+const ITERATIONS: u64 = 1_000_000_000_000_000;
 
 /// A number of bytes.
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 pub struct Bytes(pub u64);
 
 impl Bytes {
-    //- Constructors -----------------------------
-
-    /// Constructs a new `Bytes` with the supplied number of kibibytes (2¹⁰ bytes).
+    /// Returns the number of bytes in the supplied number of kibibytes (2¹⁰ bytes).
     pub fn kibibytes(kibibytes: u64) -> Self {
         Bytes(kibibytes * 1024)
     }
 
-    /// Constructs a new `Bytes` with the supplied number of mebibytes (2²⁰ bytes).
+    /// Returns the number of bytes in the supplied number of mebibytes (2²⁰ bytes).
     pub fn mebibytes(mebibytes: u64) -> Self {
         Bytes(mebibytes * 1024 * 1024)
     }
 
-    /// Constructs a new `Bytes` with the supplied number of gibibytes (2³⁰ bytes).
+    /// Returns the number of bytes in the supplied number of gibibytes (2³⁰ bytes).
     pub fn gibibytes(gibibytes: u64) -> Self {
         Bytes(gibibytes * 1024 * 1024 * 1024)
     }
 }
 
-// GeometricSequence _____________________________
-
-/// Generates unique values from a geometric sequence.
-#[derive(Copy, Clone, Debug)]
-pub struct GeometricSequence {
-    current: f64,
-    factor: f64,
-}
-
-impl GeometricSequence {
-    //- Constructors -----------------------------
-
-    /// Constructs a new `GeometricSequence`.
-    pub fn new(start: u64, factor: f64) -> Self {
-        GeometricSequence { current: start as f64, factor: factor }
-    }
-}
-
-impl Iterator for GeometricSequence {
-    type Item = u64;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let start = self.current as u64;
-        while self.current as u64 == start { self.current *= self.factor; }
-        Some(start)
-    }
-}
-
-// Measurement ___________________________________
-
-/// A measurement of the execution time of a function.
-#[derive(Copy, Clone, Debug)]
-pub struct Measurement {
-    /// The number of times the function was executed.
-    pub iterations: u64,
-    /// The amount of time that elapsed while executing the function.
-    pub elapsed: Nanoseconds<u64>,
-}
-
-// Options _______________________________________
-
-/// Benchmarking options.
+/// A set of benchmarking options.
 #[derive(Copy, Clone, Debug)]
 pub struct Options {
     factor: f64,
@@ -174,8 +115,6 @@ pub struct Options {
 }
 
 impl Options {
-    //- Consumers --------------------------------
-
     /// Sets the geometric growth factor for benchmark sample iterations.
     ///
     /// **Default:** `1.01`
@@ -184,7 +123,7 @@ impl Options {
         self
     }
 
-    /// Sets the maximum amount of memory a benchmark will allocate.
+    /// Sets the maximum amount of memory benchmarks will allocate.
     ///
     /// **Default:** `Bytes::mebibytes(512)`
     pub fn memory(mut self, memory: Bytes) -> Self {
@@ -192,7 +131,7 @@ impl Options {
         self
     }
 
-    /// Sets the maximum amount of time a benchmark will run for.
+    /// Sets the maximum amount of time benchmarks will run for.
     ///
     /// **Default:** `Duration::new(5, 0)`
     pub fn time(mut self, time: Duration) -> Self {
@@ -202,84 +141,74 @@ impl Options {
 }
 
 impl Default for Options {
-    fn default() -> Options {
-        Options { factor: 1.01, memory: Bytes::mebibytes(512), time: Duration::new(5, 0).into() }
+    fn default() -> Self {
+        let factor = 1.01;
+        let memory = Bytes::mebibytes(512);
+        let time = Duration::new(5, 0).into();
+        Options { factor, memory, time }
     }
 }
 
-//================================================
-// Functions
-//================================================
-
-fn bench_impl<F>(name: &str, f: F) where F: FnOnce() -> Vec<Measurement> {
-    let stopwatch = Stopwatch::new();
-    let measurements = f();
-    let analysis = analyze(&measurements);
-    let prefix = format!("{} ({}) ...", name, stopwatch.elapsed());
-    if measurements.len() < 2 || analysis.beta.0 < 0.0 {
-        println!("{:<32} {:>15}", prefix, "           not enough measurements");
-    } else {
-        let beta = utility::thousands(analysis.beta.0, '_');
-        println!("{:<32} {:>15} ns/iter ({:.3} R²)", prefix, beta, analysis.r2);
-    }
+/// A sample of the execution time of a function.
+#[derive(Copy, Clone, Debug)]
+pub struct Sample {
+    /// The number of times the function was executed.
+    pub iterations: u64,
+    /// The number of nanoseconds that elapsed while executing the function.
+    pub elapsed: Nanoseconds<u64>,
 }
 
-const ITERATIONS: u64 = 1_000_000_000_000_000;
-
-fn measure_impl<F>(
-    options: &Options, mut f: F
-) -> Vec<Measurement> where F: FnMut(u64) -> Option<Nanoseconds<u64>> {
-    let mut measurements = vec![];
-    let mut sequence = GeometricSequence::new(1, options.factor).take_while(|i| *i <= ITERATIONS);
-    let stopwatch = Stopwatch::new();
-    while stopwatch.elapsed() < options.time {
-        if let Some(iterations) = sequence.next() {
-            if let Some(elapsed) = f(iterations) {
-                measurements.push(Measurement { iterations: iterations, elapsed: elapsed });
-            } else {
-                break;
-            }
-        } else {
-            break;
-        }
-    }
-    measurements
+/// A statistical analysis of a set of execution time samples.
+#[derive(Copy, Clone, Debug)]
+pub struct Analysis {
+    /// The y-intercept of the simple linear regression model function.
+    pub alpha: Nanoseconds<f64>,
+    /// The slope of the simple linear regression model function.
+    pub beta: Nanoseconds<f64>,
+    /// The goodness of fit of the simple linear regression model function.
+    pub r2: f64,
 }
 
-/// Analyzes the supplied timing data and returns the resulting analysis.
-pub fn analyze(measurements: &[Measurement]) -> Analysis {
-    let data = measurements.iter()
-        .map(|m| (m.iterations as f64, m.elapsed.0 as f64))
-        .collect::<Vec<_>>();
-    let (alpha, beta, r2) = statistics::regression(&data);
-    Analysis { alpha: Nanoseconds(alpha), beta: Nanoseconds(beta), r2: r2 }
+impl Analysis {
+    /// Returns a new analysis for the supplied samples.
+    fn new(samples: &[Sample]) -> Self {
+        let Model { alpha, beta, r2 } = samples.iter()
+            .map(|m| (m.iterations as f64, m.elapsed.0 as f64))
+            .collect::<Model>();
+        Self { alpha: Nanoseconds(alpha), beta: Nanoseconds(beta), r2 }
+    }
 }
 
 /// Benchmarks the supplied function and prints the results.
-pub fn bench<T, F>(options: &Options, name: &str, f: F) where F: FnMut() -> T {
+pub fn bench<T>(options: &Options, name: &str, f: impl FnMut() -> T) {
     bench_impl(name, move || measure(options, f));
 }
 
 /// Benchmarks the supplied function ignoring drop time and prints the results.
 ///
 /// See [`measure_drop`](fn.measure_drop.html) for more information.
-pub fn bench_drop<T, F>(options: &Options, name: &str, f: F) where F: FnMut() -> T {
+pub fn bench_drop<T>(options: &Options, name: &str, f: impl FnMut() -> T) {
     bench_impl(name, move || measure_drop(options, f));
 }
 
 /// Benchmarks the supplied function ignoring setup time and prints the results.
 ///
 /// See [`measure_setup`](fn.measure_setup.html) for more information.
-pub fn bench_setup<I, S, T, F>(
-    options: &Options, name: &str, setup: S, f: F
-) where S: FnMut() -> I, F: FnMut(I) -> T {
+pub fn bench_setup<I, T>(
+    options: &Options,
+    name: &str,
+    setup: impl FnMut() -> I,
+    f: impl FnMut(I) -> T,
+) {
     bench_impl(name, move || measure_setup(options, setup, f));
 }
 
 /// Measures the execution time of the supplied function.
-pub fn measure<T, F>(options: &Options, mut f: F) -> Vec<Measurement> where F: FnMut() -> T {
+pub fn measure<T>(
+    options: &Options, mut f: impl FnMut() -> T
+) -> Vec<Sample> {
     measure_impl(options, |iterations| {
-        let stopwatch = Stopwatch::new();
+        let stopwatch = Stopwatch::default();
         for _ in 0..iterations { retain(f()); }
         Some(stopwatch.elapsed())
     })
@@ -287,21 +216,27 @@ pub fn measure<T, F>(options: &Options, mut f: F) -> Vec<Measurement> where F: F
 
 /// Measures the execution time of the supplied function ignoring drop time.
 ///
-/// This function does not include the time it takes to drop the values returned by the supplied
-/// function in the measurements. This can be useful when you want to exclude the running time of a
-/// slow implementation of `Drop` from your benchmark. However, it should be noted that this
-/// function introduces a very small amount of overhead which will be reflected in the measurements
+/// This function does not include the time it takes to drop the values returned
+/// by the supplied function in the measurements. This can be useful when you
+/// want to exclude the running time of a slow implementation of `Drop` from
+/// your benchmark. However, it should be noted that this function introduces a
+/// very small amount of overhead which will be reflected in the measurements
 /// (typically of the order of a few nanoseconds).
 ///
-/// **Warning:** This function can potentially allocate very large amounts of memory. The `memory`
-/// option controls the maximum this function is allowed to allocate.
-pub fn measure_drop<T, F>(options: &Options, mut f: F) -> Vec<Measurement> where F: FnMut() -> T {
+/// **Warning:** This function can potentially allocate very large amounts of
+/// memory. The `memory` option controls the maximum amount of memory this
+/// function is allowed to allocate.
+pub fn measure_drop<T>(
+    options: &Options, mut f: impl FnMut() -> T
+) -> Vec<Sample> {
     measure_impl(options, |iterations| {
-        if options.memory < Bytes(iterations * cmp::max(1, mem::size_of::<T>() as u64)) {
+        let size = cmp::max(1, mem::size_of::<T>() as u64);
+        if options.memory < Bytes(iterations * size) {
             return None;
         }
+
         let mut outputs = Vec::with_capacity(iterations as usize);
-        let stopwatch = Stopwatch::new();
+        let stopwatch = Stopwatch::default();
         for _ in 0..iterations { outputs.push(f()); }
         let elapsed = stopwatch.elapsed();
         mem::drop(outputs);
@@ -311,23 +246,29 @@ pub fn measure_drop<T, F>(options: &Options, mut f: F) -> Vec<Measurement> where
 
 /// Measures the execution time of the supplied function ignoring setup time.
 ///
-/// This function does not include the time it takes to execute the setup function in the
-/// measurements. This can be useful when you want to exclude the running time of some non-trivial
-/// setup which is needed for every execution of the supplied function. However, it should be noted
-/// that this function introduces a very small amount of overhead which will be reflected in the
-/// measurements (typically of the order of a few nanoseconds).
+/// This function does not include the time it takes to execute the setup
+/// function in the measurements. This can be useful when you want to exclude
+/// the running time of some non-trivial setup which is needed for every
+/// execution of the supplied function. However, it should be noted that this
+/// function introduces a very small amount of overhead which will be reflected
+/// in the measurements (typically of the order of a few nanoseconds).
 ///
-/// **Warning:** This function can potentially allocate very large amounts of memory. The `memory`
-/// option controls the maximum this function is allowed to allocate.
-pub fn measure_setup<I, S, T, F>(
-    options: &Options, mut setup: S, mut f: F
-) -> Vec<Measurement> where S: FnMut() -> I, F: FnMut(I) -> T {
+/// **Warning:** This function can potentially allocate very large amounts of
+/// memory. The `memory` option controls the maximum amount of memory this
+/// function is allowed to allocate.
+pub fn measure_setup<I, T>(
+    options: &Options,
+    mut setup: impl FnMut() -> I,
+    mut f: impl FnMut(I) -> T,
+) -> Vec<Sample> {
     measure_impl(options, |iterations| {
-        if options.memory < Bytes(iterations * cmp::max(1, mem::size_of::<I>() as u64)) {
+        let size = cmp::max(1, mem::size_of::<I>() as u64);
+        if options.memory < Bytes(iterations * size) {
             return None;
         }
+
         let inputs = retain((0..iterations).map(|_| setup()).collect::<Vec<_>>());
-        let stopwatch = Stopwatch::new();
+        let stopwatch = Stopwatch::default();
         for input in inputs { retain(f(input)); }
         Some(stopwatch.elapsed())
     })
@@ -335,9 +276,37 @@ pub fn measure_setup<I, S, T, F>(
 
 /// A function that prevents the optimizer from eliminating the supplied value.
 ///
-/// This function may not operate correctly or may have poor performance on the stable and beta
-/// channels of Rust. If you are using a nightly release of Rust, enable the `nightly` crate feature
-/// to enable a superior implementation of this function.
+/// This function may not operate correctly or may have poor performance on the
+/// stable and beta channels of Rust. If you are using a nightly release of
+/// Rust, enable the `nightly` crate feature to enable a superior implementation
+/// of this function.
 pub fn retain<T>(value: T) -> T {
-    utility::black_box(value)
+    black_box(value)
+}
+
+/// Prints an analysis of the samples produced by the supplied function.
+fn bench_impl(name: &str, f: impl FnOnce() -> Vec<Sample>) {
+    let stopwatch = Stopwatch::default();
+    let samples = f();
+    let elapsed = stopwatch.elapsed();
+    let analysis = Analysis::new(&samples);
+
+    let prefix = format!("{} ({}) ...", name, elapsed);
+    if samples.len() < 2 || analysis.beta.0 < 0.0 {
+        println!("{:<32} {:>15}", prefix, "           not enough samples");
+    } else {
+        let beta = format_number(analysis.beta.0, 3, '_');
+        println!("{:<32} {:>15} ns/iter ({:.3} R²)", prefix, beta, analysis.r2);
+    }
+}
+
+/// Collects samples produced by the supplied function.
+fn measure_impl(
+    options: &Options, mut f: impl FnMut(u64) -> Option<Nanoseconds<u64>>
+) -> Vec<Sample> {
+    let stopwatch = Stopwatch::default();
+    GeometricSequence::new(1, options.factor)
+        .take_while(|i| *i <= ITERATIONS && stopwatch.elapsed() < options.time)
+        .filter_map(|i| Some(Sample { iterations: i, elapsed: f(i)? }))
+        .collect()
 }