+++ src/time.rs
@@ -17,14 +17,8 @@
 use std::fmt;
 use std::time::{Duration, Instant};
 
-//================================================
-// Structs
-//================================================
-
-// Nanoseconds ___________________________________
-
 /// A number of nanoseconds.
-#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct Nanoseconds<T>(pub T);
 
 impl fmt::Display for Nanoseconds<u64> {
@@ -35,30 +29,25 @@ impl fmt::Display for Nanoseconds<u64> {
 
 impl From<Duration> for Nanoseconds<u64> {
     fn from(duration: Duration) -> Nanoseconds<u64> {
-        Nanoseconds((duration.as_secs() * 1_000_000_000) + duration.subsec_nanos() as u64)
+        let nanos = u64::from(duration.subsec_nanos());
+        Nanoseconds((duration.as_secs() * 1_000_000_000) + nanos)
     }
 }
-
-// Stopwatch _____________________________________
 
 /// A high-precision stopwatch.
 #[derive(Clone, Copy, Debug)]
-pub struct Stopwatch {
-    start: Instant,
-}
+pub struct Stopwatch(Instant);
 
 impl Stopwatch {
-    //- Constructors -----------------------------
-
-    /// Constructs a new `Stopwatch`.
-    pub fn new() -> Self {
-        Stopwatch { start: Instant::now() }
+    /// Returns the number of nanoseconds that have elapsed since this stopwatch
+    /// was last constructed or reset.
+    pub fn elapsed(self) -> Nanoseconds<u64> {
+        (Instant::now() - self.0).into()
     }
+}
 
-    //- Accessors --------------------------------
-
-    /// Returns the elapsed nanoseconds since this stopwatch was constructed or reset.
-    pub fn elapsed(&self) -> Nanoseconds<u64> {
-        (Instant::now() - self.start).into()
+impl Default for Stopwatch {
+    fn default() -> Self {
+        Stopwatch(Instant::now())
     }
 }